Workable version
This is a workable version but need to be modified as you need.
Wifi and BLE will be initialized in the very beginning. 
You need to turn off one of them to make the other work.
Otherwise, ESP32 will be stuck. 

And there is a big bug cannot be fixed - ble cannot work when you turn off and turn on again. 
Still don't know what's wrong there, but I think the problem may count on N. Kolban's API.(need to be checked)